refactor : 불필요한 리렌더링 방지
